finish ninja crud app

+++ ninjanet_mern/workout_ui/src/components/WorkoutDetails.jsx
@@ -1,19 +1,24 @@
 /* eslint-disable react/prop-types */
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
+import { useAuthContext } from "../hooks/useAuthContext";
+
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
 
 const WORKOUTS_API_URL = "http://localhost:5174/api/workouts/";
 
 function WorkoutDetails({ workout }) {
   const { dispatch } = useWorkoutsContext();
+  const { user } = useAuthContext();
 
   const handleDelete = async () => {
+    if (!user) return;
+
     const response = await fetch(WORKOUTS_API_URL + workout._id, {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${user.token}`,
+      },
     });
-
-    // if (!response.ok) {
-    // }
 
     if (response.ok) {
       dispatch({ type: "DELETE_WORKOUT", payload: workout._id });
